==> notebooks/2-filter-scansion-info.ipynb <==
# Import libraries

import os
import pandas as pd
from argparse import Namespace
from pathlib import Path
import json
import ast

args = Namespace(
    filename= "rantanplan-rhyme-stress-dataset",
    source_dir = Path(os.path.abspath(os.path.join("..", "rantanplan-data", "plain-lyrics"))),
    save_dir = Path(os.path.abspath(os.path.join("..", "rantanplan-data", "rhyme-stress"))),
)

# Helper functions

def get_rhyme(scansion):
    """
    Returns a string containing the rhyme pattern from scansion list of lines.
    """
    if scansion:
        sc = ast.literal_eval(scansion)
        return "".join([line["rhyme"] for line in sc])
    else:
        return None

def get_stress(scansion):
    """
    Returns a string containing the rhyme pattern from scansion list of lines.
    """
    if scansion:
        sc = ast.literal_eval(scansion)
        return [line["rhythm"]['stress'] for line in sc]
    else:
        return None

# Read all files in `source_dir`, extract rhyme and stress and save to a file

%%time
file_list = []
for path in args.source_dir.glob("*.parquet"):
    df = pd.read_parquet(path)
    df['rhyme'] = [get_rhyme(sc) for sc in df['scansion']]
    df['stress'] = [get_stress(sc) for sc in df['scansion']]
    file_list.append(df[["author", "title", "text", "id", "rhyme", "stress"]])
total = pd.concat(file_list)

total.sort_values(by=['id'], inplace=True)

total.to_parquet(args.save_dir / f"{args.filename}.parquet",compression="gzip")

total.head(10)
