==> notebooks/3-rhyme-groups.ipynb <==
- comprobar monorrimo (a+)
- comprobar sin rima (\-+)
- dividir cadena rima en grupos de 4
- por cada grupo, comprobar:
    + enclosed rhyme - rima abrazada abba 
    + cross - rima cruzada abab
- dividir en grupos de 2    
    - pair - rima pareada aa
- contar porcentaje de cada tipo
- definir tipos de canciones por rima


import re
import pandas as pd
from collections import Counter

MONO_RE = re.compile(r"a+", re.VERBOSE | re.I) #aaaa
PAIR_RE = re.compile(r"([a-zñ])\1{1,}", re.VERBOSE | re.I) #aa
CROSS_RE = re.compile(r"([a-zñ])(?!\1)([a-zñ])\1\2", re.VERBOSE | re.I) # abab
ENCLOSED_RE = re.compile(r"([a-zñ])(?!\1)([a-zñ])\2\1", re.VERBOSE | re.I) #abba
NO_RHYME_RE = re.compile(r"\-+", re.VERBOSE | re.I) #----

def calc_percents(rhyme:str, counts: dict) -> dict:
    total = len(rhyme)
    percents = {} 
    for key, value in counts.items():
        percents[key] = (value/total) * 100 if total>0 else 0
    return percents

def split_and_count(rhyme, step, regex): 
    count=0
    for i in range(0, len(rhyme), step):
        chunk = rhyme[i:i+step]
        if re.fullmatch(regex, chunk):
            count += step 
    return count

def count_couplet(rhyme):
    count=0
    for match in re.finditer(PAIR_RE, rhyme):
        span_len = match.end() - match.start()
        if span_len % 2 == 0:
            count += span_len
    return count

def count_rhyme(rhyme: str):
    counts = {"monorhyme": 0,
             "crossed_rhyme": 0,
             "enclosed_rhyme": 0,
             "couplet": 0,
             "no_rhyme": 0}
    is_mono = re.fullmatch(MONO_RE, rhyme)
    if is_mono:
        counts["monorhyme"] = len(rhyme)
    else:
        counts["no_rhyme"] = Counter(rhyme)['-']
        counts["enclosed_rhyme"] = split_and_count(rhyme, 4, ENCLOSED_RE)
        counts["crossed_rhyme"] = split_and_count(rhyme, 4, CROSS_RE)
        if counts["enclosed_rhyme"] == 0 and counts["crossed_rhyme"] == 0:
            counts["couplet"] = count_couplet(rhyme)
    percents = calc_percents(rhyme, counts)
    return percents

df = pd.read_parquet("../rantanplan-data/rhyme-stress/spotify-rantanplan-lyrics.parquet")

df = df.loc[df["spotify_id"].notnull()]

df = df.reset_index(drop=True)

counts_rhyme = [count_rhyme(x) for x in df["rhyme"]]

total = df.join(pd.json_normalize(counts_rhyme))

total.to_parquet("../spotify-rhyme-groups.parquet", compression="gzip", index=False)

total.head()
