==> notebooks/4-lenght-groups.ipynb <==
Propuesta de intervalos por número de sílabas:
- Versos con menos de 8 sílabas
- Versos con 8 sílabas
- Versos con 9 y 10 sílabas
- Versos con 11 sílabas
- Versos con más de 11 sílabas


import re
import pandas as pd

def calc_percents(stress:list, counts: dict) -> dict:
    total = len(stress)
    percents = {} 
    for key, value in counts.items():
        percents[key] = (value/total) * 100 if total>0 else 0
    return percents

def count_lengths(stress: list):
    counts = {"arte_menor": 0,
             "octosílabo": 0,
             "enea_decasílabo": 0,
             "endecasílabo": 0,
             "arte_mayor":0}
    for pattern in stress:
        length = len(pattern)
        #comprobar menor
        if length < 8:
            counts["arte_menor"] += 1
        elif length == 8:
            counts["octosílabo"] += 1
        elif length < 10:
            counts["enea_decasílabo"] += 1
        elif length == 11:
            counts["endecasílabo"] += 1
        elif length > 11:
            counts["arte_mayor"] += 1
    percents = calc_percents(stress, counts)
    return percents

df = pd.read_parquet("../rantanplan-data/rhyme-stress/spotify-rantanplan-lyrics.parquet")

df = df.loc[df["spotify_id"].notnull()]

df = df.reset_index(drop=True)

count_list = [count_lengths(x) for x in df["stress"]]

total = df.join(pd.json_normalize(count_list))

total.to_parquet("../spotify-lengths-groups.parquet", compression="gzip", index=False)

total.head()
