==> notebooks/5-merge.ipynb <==
import pandas as pd
from pathlib import Path
import ast

# fingerprint

rhyme_groups = pd.read_parquet("../spotify-rhyme-groups.parquet",
                               columns=["spotify_id", "rhyme", "monorhyme", "crossed_rhyme", "enclosed_rhyme",
                                       "couplet", "no_rhyme"])

length_groups = pd.read_parquet("../spotify-lengths-groups.parquet",
                                columns=["spotify_id", "stress", "arte_menor", "octosílabo", "enea_decasílabo",
                                       "endecasílabo", "arte_mayor"])

spotify = pd.read_parquet("../spotify_corpus.parquet")

spotify_rhyme = spotify.merge(rhyme_groups, validate="1:1", on="spotify_id")

spotify_rhyme.head()

spotify_rhyme_lenghts = spotify_rhyme.merge(length_groups, validate="1:1", on="spotify_id")

spotify_rhyme_lenghts["n_verses"] = pd.to_numeric(spotify_rhyme_lenghts["n_verses"], downcast="integer")

def change_type(lengths, _type):
    """Change type for each element of the list"""
    new = []
    for index, element in enumerate(lengths):
        new.append(_type(element))
    return new

spotify_rhyme_lenghts["lengths"] = spotify_rhyme_lenghts["lengths"].apply(lambda x: change_type(x, int))

spotify_rhyme_lenghts["stress"] = spotify_rhyme_lenghts["stress"].apply(lambda x: change_type(x, str))

spotify_rhyme_lenghts.to_parquet("spotify_corpus.parquet", compression="gzip", index=False)

lengths_spotify = lengths_spotify.loc[lengths_spotify["spotify_id"].notnull()]

n_verses = n_verses.loc[n_verses["spotify_id"].notnull()]

duplicados = [x for x, y in enumerate(lengths_spotify["spotify_id"].duplicated()) if y]

duplicados

merged = lengths_spotify.merge(n_verses, validate="1:1")

merged.head()

spotify_merged = spotify.merge(merged, validate="1:1")

lista = spotify_merged["lengths"][0]

spotify_merged['lengths'].apply(ast.literal_eval)

spotify_merged.head(2)

spotify_merged["id"] = pd.to_numeric(spotify_merged["id"], downcast="integer")
spotify_merged["artist_name"] = spotify_merged["artist_name"].astype("category")

genres = genres=pd.read_csv("../spotify_genres.csv")

genres = genres.loc[genres["spotify_id"].notnull()]

spotify_total = spotify_merged.merge(genres[["spotify_id", "genres"]], validate="1:1")

spotify_total.shape

spotify_total = spotify_total.loc[spotify_total["lyrics"].notnull()]

spotify_total['lengths'] = spotify_total['lengths'].apply(ast.literal_eval)

spotify_total.to_parquet("spotify_corpus.parquet", compression="gzip", index=False)

df = pd.read_parquet("../spotify_corpus.parquet")

type(df["lengths"][0])

### LastFM

lastfm = pd.read_parquet("../lastfm_corpus.parquet")

df = pd.read_parquet("../spotify_corpus.parquet", columns=["spotify_id",
    "rhyme", "monorhyme", "crossed_rhyme", "enclosed_rhyme", "couplet", "no_rhyme",
    "stress", "arte_menor", "octosílabo", "enea_decasílabo","endecasílabo", "arte_mayor"
])

lastfm.columns

lastfm = lastfm.loc[lastfm["spotify_id"].notnull()]

lastfm["id"] = pd.to_numeric(lastfm["id"], downcast="integer")
lastfm["artist_name"] = lastfm["artist_name"].astype("category")

lastfm_total = lastfm.merge(df, validate="1:1", on="spotify_id")

lastfm_total.shape

lastfm_total = lastfm_total.loc[lastfm_total["lyrics"].notnull()]

lastfm_total["lengths"] = lastfm_total["lengths"].apply(lambda x: change_type(x, int))

lastfm_total["stress"] = lastfm_total["stress"].apply(lambda x: change_type(x, str))

lastfm_total.shape

lastfm_total.to_parquet("lyrics_digital_fingerprint.parquet", compression="gzip", index=False)

lastfm_total.columns
