==> notebooks/6-subset.ipynb <==
import pandas as pd
import re

stopwords = {'estribillo','x2','x3','x4','x5', 'coro-estribillo','(x2)','(x3)','(x4)', '()', '(repite estribillo 3 veces)'}

REGEX_PAT = re.compile(r"\b(?:" + ("|").join(stopwords) + ")\\b", re.IGNORECASE)

df = pd.read_parquet("./digital_identity_token_dataset.parquet", columns=['spotify_id', 'lyrics'])

subset = df.sample(5000, random_state=5)[["spotify_id", "lyrics"]]

subset.loc[subset["lyrics"].str.contains("estribillo")]

subset["lyrics"] = subset["lyrics"].str.replace(REGEX_PAT, "", regex=True)

subset.loc[subset["lyrics"].str.contains("x3")]

subset.to_csv("subset.csv", index=False)
