==> notebooks/7-LDA_no_stopwords_no_lemmas.ipynb <==
from gensim.corpora.dictionary import Dictionary
from gensim.models.ldamodel import LdaModel
from gensim.test.utils import datapath
import pandas as pd
import spacy
from spacy.language import Language

stopword_list = []

with open("vacias.txt", 'r') as f:
    stopword_list = f.read().split("\n")

# print(stopword_list)

@Language.component("remove_stopwords")
def remove_stopwords(doc):
    doc = [token for token in doc if not token.is_stop]
    return doc

@Language.component("remove_punct")
def remove_punct(doc):
    doc = [token for token in doc if not token.is_punct]
    return doc

@Language.component("remove_spaces")
def remove_spaces(doc):
    doc = [token for token in doc if not token.is_space]
    return doc

@Language.component("get_text")
def get_text(doc):
    doc = [token.text for token in doc]
    return doc

nlp = spacy.load('es_core_news_lg', exclude=["ner"])

print(len(nlp.Defaults.stop_words))

nlp.Defaults.stop_words |= set(stopword_list)

for key in stopword_list:
    nlp.vocab[key].is_stop = True

print(len(nlp.Defaults.stop_words))

rm_stopwords = nlp.add_pipe("remove_stopwords", name="rm_stopwords", last=True)
rm_punct = nlp.add_pipe("remove_punct", name="rm_punct", last=True)
rm_spaces = nlp.add_pipe("remove_spaces", name="rm_spaces", last=True)
get_text = nlp.add_pipe("get_text", name="get_text", last=True)

df = pd.read_parquet("./digital_identity_token_dataset.parquet", columns=['spotify_id', 'lyrics'])

df["doc"] = list(nlp.pipe(df["lyrics"]))

dictionary = Dictionary(df["doc"])

corpus = [dictionary.doc2bow(text) for text in df["doc"]]

lda = LdaModel(corpus=corpus, num_topics=10, id2word=dictionary)

model_name = "./ldamodels/lda_text_no_stopwords"

lda.save(model_name)

lda = LdaModel.load(model_name)
dictionary = Dictionary.load(f"{model_name}.id2word")
